Struct near_clp::NearCLP

pub struct NearCLP {
    pub fee_dst: AccountId,
    pub owner: AccountId,
    // some fields omitted
}

NearCLP is the main contract for managing the swap pools and liquidity. It implements the NEARswap functionality.

Fields
fee_dst: AccountIdowner: AccountId
Implementations

impl NearCLP

pub fn new(owner: AccountId) -> Self

pub fn set_fee_dst(&mut self, fee_dst: AccountId)

pub fn change_owner(&mut self, new_owner: AccountId)

Owner is an account (can be a multisig) who has management rights to update fee size.

pub fn create_pool(&mut self, token: AccountId)

Allows any user to creat a new near-token pool. Each pool is identified by the token account - which we call the Pool Reserve Token. If a pool for give token exists then "E1" assert exception is thrown. TODO: charge user for a storage created!

pub fn pool_info(&self, token: &AccountId) -> Option<PoolInfo>

Extracts public information of the token pool.

pub fn list_pools(&self) -> Vec<AccountId>

Returns list of pools identified as their reserve token AccountId.

pub fn add_liquidity(
&mut self,
token: AccountId,
max_token_amount: Balance,
min_shares_amount: Balance
)

Increases Near and the Reserve token liquidity. The supplied funds must preserver current ratio of the liquidity pool.

pub fn withdraw_liquidity(
&mut self,
token: AccountId,
shares: Balance,
min_near: Balance,
min_tokens: Balance
)

Redeems shares for liquidity stored in this pool with condition of getting at least min_near of Near and min_tokens of reserve. Shares are note exchengable between different pools.

pub fn shares_balance_of(&self, token: AccountId, owner: AccountId) -> Balance

Returns the owner balance of shares of a pool identified by token.

pub fn swap_near_to_reserve_exact_in(
&mut self,
token: AccountId,
min_tokens: Balance
)

Swaps NEAR to token and transfers the reserve tokens to the caller. Caller attaches near tokens he wants to swap to the transacion under a condition of receving at least min_tokens.

pub fn swap_near_to_reserve_exact_in_xfr(
&mut self,
token: AccountId,
min_tokens: Balance,
recipient: AccountId
)

Same as swap_near_to_reserve_exact_in, but user additionly specifies the recipient who will receive the tokens after the swap.

pub fn swap_near_to_reserve_exact_out(
&mut self,
token: AccountId,
tokens_out: Balance
)

Swaps NEAR to token and transfers the reserve tokens to the caller. Caller attaches maximum amount of NEAR he is willing to swap to receive tokens_out of token wants to swap to the transacion. Surplus of NEAR tokens will be returned. Transaction will panic if the caller doesn't attach enough NEAR tokens.

pub fn swap_near_to_reserve_exact_out_xfr(
&mut self,
token: AccountId,
tokens_out: Balance,
recipient: AccountId
)

Same as swap_near_to_reserve_exact_out, but user additionly specifies the recipient who will receive the tokens after the swap.

pub fn swap_reserve_to_near_exact_in(
&mut self,
token: AccountId,
tokens_paid: Balance,
min_near: Balance
)

Swaps token to NEAR and transfers NEAR to the caller under a condition of receving at least min_near. Preceeding to this transaction, caller has to create sufficient allowance of token for this contract. TODO: Transaction will panic if a caller doesn't provide enough allowance.

pub fn swap_reserve_to_near_exact_in_xfr(
&mut self,
token: AccountId,
tokens_paid: Balance,
min_near: Balance,
recipient: AccountId
)

Same as swap_reserve_to_near_exact_in, but user additionly specifies the recipient who will receive the tokens after the swap.

pub fn swap_reserve_to_near_exact_out(
&mut self,
token: AccountId,
near_out: Balance,
max_tokens: Balance
)

Swaps token to NEAR and transfers NEAR to the caller. Caller defines the amount of NEAR he wants to receive under a condition of not spending more than max_tokens of token. Preceeding to this transaction, caller has to create sufficient allowance of token for this contract. TODO: Transaction will panic if a caller doesn't provide enough allowance.

pub fn swap_reserve_to_near_exact_out_xfr(
&mut self,
token: AccountId,
near_out: Balance,
max_tokens: Balance,
recipient: AccountId
)

Same as swap_reserve_to_near_exact_out, but user additionly specifies the recipient who will receive the tokens after the swap.

pub fn swap_tokens_exact_in(
&mut self,
from: AccountId,
to: AccountId,
tokens_from: Balance,
min_tokens_to: Balance
)

Swaps two different tokens. Caller defines the amount of tokens he wants to swap under a condition of receving at least min_tokens_to. Preceeding to this transaction, caller has to create sufficient allowance of token_from for this contract.

pub fn swap_tokens_exact_in_xfr(
&mut self,
from: AccountId,
to: AccountId,
tokens_from: Balance,
min_tokens_to: Balance,
recipient: AccountId
)

Same as swap_tokens_exact_in, but user additionly specifies the recipient who will receive the tokens after the swap.

pub fn swap_tokens_exact_out(
&mut self,
from: AccountId,
to: AccountId,
tokens_to: Balance,
max_tokens_from: Balance
)

Swaps two different tokens. Caller defines the amount of tokens he wants to receive under a of not spending more than max_tokens_from. Preceeding to this transaction, caller has to create sufficient allowance of token_from for this contract.

pub fn swap_tokens_exact_out_xfr(
&mut self,
from: AccountId,
to: AccountId,
tokens_to: Balance,
max_tokens_from: Balance,
recipient: AccountId
)

Same as swap_tokens_exact_out, but user additionly specifies the recipient who will receive the tokens after the swap.

pub fn price_near_to_token_in(
&self,
token: AccountId,
near_in: Balance
) -> Balance

Calculates amount of tokens user will recieve when swapping near_in for token assets

pub fn price_near_to_token_out(
&self,
token: AccountId,
tokens_out: Balance
) -> Balance

Calculates amount of NEAR user will need to swap if he wants to receive tokens_out of tokens

pub fn price_token_to_near_in(
&self,
token: AccountId,
tokens_in: Balance
) -> Balance

Calculates amount of NEAR user will recieve when swapping tokens_in for NEAR.

pub fn price_token_to_near_out(
&self,
token: AccountId,
near_out: Balance
) -> Balance

Calculates amount of tokens user will need to swap if he wants to receive tokens_out of tokens

pub fn price_token_to_token_in(
&self,
from: AccountId,
to: AccountId,
tokens_in: Balance
) -> Balance

Calculates amount of tokens to user will receive when swapping tokens_in of from

pub fn price_token_to_token_out(
&self,
from: AccountId,
to: AccountId,
tokens_out: Balance
) -> Balance

Calculates amount of tokens from user will need to swap if he wants to receive tokens_out of tokens to

pub fn add_liquidity_transfer_callback(&mut self, token: AccountId)